% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_G_function.R
\name{create_G_function}
\alias{create_G_function}
\title{Create Step Function for Estimated Survival (G)}
\usage{
create_G_function(unique_times, G_vals)
}
\arguments{
\item{unique_times}{A numeric vector of increasing unique times (e.g., censoring times).}

\item{G_vals}{A numeric vector of the same length as \code{unique_times}, corresponding to estimated values of the survival function at each time.}
}
\value{
A function \code{G(t)} that can be evaluated on a numeric vector of time points.
}
\description{
This function creates a right-continuous step function (vectorized) based on
estimated values of the censoring survival function at given unique times.
}
\details{
The resulting function can be used to evaluate \eqn{G(t)} for any vector of time points.
}
\examples{
unique_times <- c(1, 2, 3, 5)
G_vals <- c(1.0, 0.9, 0.85, 0.75)
G_fun <- create_G_function(unique_times, G_vals)
G_fun(c(0, 2, 4, 6))  # Returns: 1.0, 0.9, 0.85, 0.75

}
