% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ml_Cox.R
\name{Ml_Cox}
\alias{Ml_Cox}
\title{Maximum Likelihood Estimation for the Cox Model (Optimized)}
\usage{
Ml_Cox(data, max_iter = 100, tol = 1e-06)
}
\arguments{
\item{data}{A data frame with columns: \code{times}, \code{status}, and covariates starting from the 4th column.}

\item{max_iter}{Maximum number of iterations (default = 100).}

\item{tol}{Convergence tolerance for the Euclidean norm (default = 1e-6).}
}
\value{
A named list with one element:
\describe{
\item{beta}{A numeric vector of estimated regression coefficients in the
Cox proportional hazards model. Each coefficient represents the log hazard
ratio associated with the corresponding covariate.}
}
}
\description{
This function computes the maximum likelihood estimates of the regression coefficients in the Cox proportional hazards model using Newton-Raphson iterations.
}
