% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_format.R
\name{check_data_format}
\alias{check_data_format}
\title{Check the Format of a Data Frame for Multicentre Competing Risks with Frailty}
\usage{
check_data_format(df)
}
\arguments{
\item{df}{A data frame to be checked.}
}
\value{
Returns \code{TRUE} if the data frame is correctly structured; otherwise, returns
a character string describing the cause of failure.
}
\description{
Validates whether the submitted data frame is correctly structured for the analysis
of multicentre competing risks with frailty.
}
\examples{
\dontrun{
n_cov = 2
n_cluster = 5
n_per_cluster = 100
n = n_cluster * n_per_cluster
a1 = 0
b1 = 1
G <- rep(1:n_cluster, each = n_per_cluster)
Z <- matrix(runif(n * n_cov, a1, b1), ncol = n_cov)
df = simulate_data(
  G,
  Z,
  prop = 0.6,
  beta = c(0),
  theta = 0.01,
  cens = TRUE,
  pcens = 0.25,
  tau = 0
)
check_data_format(df)
}

}
